separated initalize logic and App
